medical-telegram-warehouse/requirements.txt
# FastAPI and web server
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# dbt
dbt-core==1.7.0
dbt-postgres==1.7.0

# Telegram
telethon==1.32.1

# Data processing
pandas==2.1.3
numpy==1.26.2

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.1

# Utilities
python-dotenv==1.0.0
python-multipart==0.0.6

# Orchestration
dagster==1.6.0
dagster-postgres==0.23.0
dagster-webserver==1.6.0

# YOLO and Computer Vision
ultralytics==8.1.0
opencv-python==4.8.1.78
pillow==10.1.0